Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    long double a;
+    double a;
     scanf("%ld",&a);
     printf("You entered: %ld",a);
     return 0;
